src/VisionService/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
opencv-python-headless==4.10.0.84
numpy==2.0.2
pillow==10.4.0
chess==1.10.0
